+++ quartz-js/components/load_more/load_more.state.js
@@ -1,0 +1,4 @@
+vhxm.components.shared.load_more.state = function() {
+  this.in_view = m.prop(false);
+  this.is_loading = m.prop(false);
+};